Gather feature filtering into one module
