+++ spdivik/predefined.py
@@ -25,6 +25,7 @@ from tqdm import tqdm
 
 import spdivik.distance as dst
 import spdivik.divik as dv
+import spdivik.feature_selection
 import spdivik.feature_selection as fs
 import spdivik.kmeans as km
 import spdivik.score as sc
@@ -49,14 +50,14 @@ def _dunn_optimized_kmeans(distance: dst.DistanceMetric,
     return best_kmeans_with_dunn
 
 
-_AMPLITUDE_FILTER = dv.FilteringMethod(
+_AMPLITUDE_FILTER = spdivik.feature_selection.FilteringMethod(
     'amplitude',
     partial(fs.select_by,
             statistic=fs.amplitude,
             discard_up_to=1,
             # discarding only lowest component, if possible
             preserve_topmost=True))
-_VARIANCE_FILTER = dv.FilteringMethod(
+_VARIANCE_FILTER = spdivik.feature_selection.FilteringMethod(
     'variance',
     partial(fs.select_by,
             statistic=fs.variance,
@@ -66,7 +67,7 @@ _VARIANCE_FILTER = dv.FilteringMethod(
 
 
 class _PrefilteringWrapper:
-    def __init__(self, prefilter: dv.FilteringMethod, divik: Divik):
+    def __init__(self, prefilter: spdivik.feature_selection.FilteringMethod, divik: Divik):
         self._prefilter = prefilter
         self._divik = divik
