+++ test/test_feature_selection.py
@@ -1,0 +1,17 @@
+import unittest
+
+
+class SelectSequentiallyTest(unittest.TestCase):
+    def test_applies_filters_on_smaller_and_smaller_data(self):
+        pass
+
+    def test_keeps_minimal_number_of_features(self):
+        pass
+
+
+class SelectSimultaneouslyTest(unittest.TestCase):
+    def test_applies_filters_on_whole_data(self):
+        pass
+
+    def test_keeps_minimal_number_of_features(self):
+        pass
